refactor(rule): split cross-validation comparators spec into domain-specific files
